+++ src/vector.rs
@@ -15,6 +15,7 @@ use crate::iter::FromNonEmptyIterator;
 use crate::iter::IntoNonEmptyIterator;
 use crate::iter::NonEmptyIterator;
 use crate::slice::NEChunks;
+use crate::Singleton;
 
 /// Like the [`vec!`] macro, but enforces at least one argument. A nice
 /// short-hand for constructing [`NEVec`] values.
@@ -1189,6 +1190,20 @@ impl<T> Extend<T> for NEVec<T> {
         I: IntoIterator<Item = T>,
     {
         self.inner.extend(iter);
+    }
+}
+
+impl<T> Singleton for NEVec<T> {
+    type Item = T;
+
+    /// ```
+    /// use nonempty_collections::{NEVec, Singleton, nev};
+    ///
+    /// let v = NEVec::singleton(1);
+    /// assert_eq!(nev![1], v);
+    /// ```
+    fn singleton(item: T) -> NEVec<T> {
+        NEVec::new(item)
     }
 }