+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # `nonempty-collections`
 
+## Unreleased
+
+#### Added
+
+- The `Singleton` trait.
+
 ## 1.0.1 (2025-09-06)
 
 #### Added